requirements.txt
numpy~=1.23.5
scipy~=1.10.0
matplotlib~=3.7.0
pandas~=1.5.3
setuptools~=65.6.3
sympy~=1.11.1
scikit-learn~=1.2.2
pillow~=9.4.0
python-dateutil~=2.8.2